Add and wire CheckZMFilter in SqlSeverBytesWriter
The filter checks if a geometry can carry z- and/or m-ordinate values and sets emitZ and emitM values accordingly based of _emitZ and _emitM.

This removes the necessity of filling and removing values from ZValues and MValues.

The drawback is, that the default value for WKTReader.IsOldNtsCoordinateSyntaxAllowed is true always creates sequences with a dimension of 3 (=>can carry z), even though these values might all be Double.NaN. Therefor I had to adjust the unit tests.