Provide a custom way to validate the geometry when written to SQL. (#25)
This allows to override the default IsValid of the NTS Geometry and provide an own strategy. This way you can use the IsValidOp with the option SelfTouchingRingFormingHoleValid flagged. Or just make them all valid following the NTS way. Either way the developer will have more control, and if we would manage to provide a IsValidSqlServerOp, it could be passed this way.

relates to #12